requirements.txt
Pillow==5.0.0
requests==2.18.4
bs4
wheel
lxml
baidu-aip==2.0.0.1